# Performance improvements to vectorized Span.Reverse