See original PR: #70944
Reverted: #78605

I reverted the overlapping for int and long and and chose to keep it in-between for byte/char where it still shows an improvement to overlap.

Machine:
``` ini
BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.22621.819)
AMD Ryzen 7 3700X, 1 CPU, 16 logical and 8 physical cores
.NET SDK=7.0.100
  [Host]     : .NET 7.0.0 (7.0.22.51805), X64 RyuJIT AVX2
  Job-AORABE : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2
  Job-BUGVTU : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2
```

### Byte 
```csharp
[HideColumns("Job", "Error", "StdDev", "Median", "RatioSD")]
public class ReverseBytes
{
  [Params(16, 24, 32, 36, 38, 40, 42, 44, 48, 56, 64)]
  public int NumberOfValues { get; set; }

  private byte[] _values;

  [GlobalSetup]
  public void Setup()
  {
      _values = ValuesGenerator.Array<byte>(NumberOfValues);
  }

  [Benchmark]
  public void Reverse() => _values.AsSpan().Reverse();
}
```

|  Method |                                                                                                          Toolchain | NumberOfValues |      Mean | Ratio |
|-------- |------------------------------------------------------------------------------------------------------------------- |--------------- |----------:|------:|
| **Reverse** | **main** |             **16** |  **7.880 ns** |  **1.00** |
| Reverse |   PR |             16 |  2.197 ns |  0.28 |
|         |                                                                                                                    |                |           |       |
| **Reverse** | **main** |             **24** | **10.383 ns** |  **1.00** |
| Reverse |   PR |             24 |  2.772 ns |  0.27 |
|         |                                                                                                                    |                |           |       |
| **Reverse** | **main** |             **32** |  **3.795 ns** |  **1.00** |
| Reverse |   PR |             32 |  2.852 ns |  0.75 |
|         |                                                                                                                    |                |           |       |
| **Reverse** | **main** |             **36** |  **4.753 ns** |  **1.00** |
| Reverse |   PR |             36 |  3.373 ns |  0.71 |
|         |                                                                                                                    |                |           |       |
| **Reverse** | **main** |             **38** |  **5.328 ns** |  **1.00** |
| Reverse |   PR |             38 |  3.985 ns |  0.75 |
|         |                                                                                                                    |                |           |       |
| **Reverse** | **main** |             **40** |  **6.299 ns** |  **1.00** |
| Reverse |   PR |             40 |  3.387 ns |  0.54 |
|         |                                                                                                                    |                |           |       |
| **Reverse** | **main** |             **42** |  **6.714 ns** |  **1.00** |
| Reverse |   PR |             42 |  4.236 ns |  0.63 |
|         |                                                                                                                    |                |           |       |
| **Reverse** | **main** |             **44** |  **7.172 ns** |  **1.00** |
| Reverse |   PR |             44 |  3.479 ns |  0.49 |
|         |                                                                                                                    |                |           |       |
| **Reverse** | **main** |             **48** |  **8.631 ns** |  **1.00** |
| Reverse |   PR |             48 |  4.737 ns |  0.55 |
|         |                                                                                                                    |                |           |       |
| **Reverse** | **main** |             **56** | **10.799 ns** |  **1.00** |
| Reverse |   PR |             56 |  4.616 ns |  0.43 |
|         |                                                                                                                    |                |           |       |
| **Reverse** | **main** |             **64** |  **4.063 ns** |  **1.00** |
| Reverse |   PR |             64 |  3.028 ns |  0.75 |

### Char
```csharp
[HideColumns("Job", "Error", "StdDev", "Median", "RatioSD")]
public class ReverseChars
{
  [Params(8, 12, 16, 20, 24, 28, 32, 40, 48)]
  public int NumberOfValues { get; set; }

  private char[] _values;

  [GlobalSetup]
  public void Setup()
  {
      // taken from dotnet/performance
      _values = ValuesGenerator.Array<char>(NumberOfValues);
  }

  [Benchmark]
  public void Reverse() => _values.AsSpan().Reverse();
}
```

|  Method |                                                                                                          Toolchain | NumberOfValues |     Mean | Ratio |
|-------- |------------------------------------------------------------------------------------------------------------------- |--------------- |---------:|------:|
| **Reverse** | **main** |              **8** | **3.918 ns** |  **1.00** |
| Reverse |   PR |              8 | 4.145 ns |  1.06 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **12** | **5.724 ns** |  **1.00** |
| Reverse |   PR |             12 | 6.470 ns |  1.13 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **16** | **3.239 ns** |  **1.00** |
| Reverse |   PR |             16 | 2.786 ns |  0.86 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **20** | **4.361 ns** |  **1.00** |
| Reverse |   PR |             20 | 4.111 ns |  0.94 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **24** | **5.996 ns** |  **1.00** |
| Reverse |   PR |             24 | 4.637 ns |  0.77 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **28** | **8.948 ns** |  **1.00** |
| Reverse |   PR |             28 | 4.648 ns |  0.52 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **32** | **3.479 ns** |  **1.00** |
| Reverse |   PR |             32 | 2.518 ns |  0.72 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **40** | **5.494 ns** |  **1.00** |
| Reverse |   PR |             40 | 5.262 ns |  0.96 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **48** | **9.208 ns** |  **1.00** |
| Reverse |   PR |             48 | 3.624 ns |  0.39 |

### Int
|  Method |                                                                                                          Toolchain | NumberOfValues |     Mean | Ratio |
|-------- |------------------------------------------------------------------------------------------------------------------- |--------------- |---------:|------:|
| **Reverse** | **main** |              **4** | **2.692 ns** |  **1.00** |
| Reverse |   PR |              4 | 1.876 ns |  0.70 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |              **8** | **3.325 ns** |  **1.00** |
| Reverse |   PR |              8 | 2.131 ns |  0.64 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **12** | **4.020 ns** |  **1.00** |
| Reverse |   PR |             12 | 3.014 ns |  0.75 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **16** | **4.275 ns** |  **1.00** |
| Reverse |   PR |             16 | 3.018 ns |  0.71 |

### Long
|  Method |                                                                                                          Toolchain | NumberOfValues |     Mean | Ratio |
|-------- |------------------------------------------------------------------------------------------------------------------- |--------------- |---------:|------:|
| **Reverse** | **main** |              **4** | **2.393 ns** |  **1.00** |
| Reverse |   PR |              4 | 2.673 ns |  1.12 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |              **8** | **2.273 ns** |  **1.00** |
| Reverse |   PR |              8 | 2.240 ns |  0.98 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **12** | **3.605 ns** |  **1.00** |
| Reverse |   PR |             12 | 2.890 ns |  0.80 |
|         |                                                                                                                    |                |          |       |
| **Reverse** | **main** |             **16** | **2.718 ns** |  **1.00** |
| Reverse |   PR |             16 | 2.708 ns |  1.00 |

@adamsitnik can you confirm if you get similar numbers on your benchmark? I think earlier we saw some slight difference between our numbers last time